Move all ugrid.metadata into common.metadata.
Removed obsolete tests.unit.metadata
