--- lib/iris/tests/unit/ugrid/metadata/__init__.py
@@ -1,7 +1,0 @@
-# Copyright Iris contributors
-#
-# This file is part of Iris and is released under the BSD license.
-# See LICENSE in the root of the repository for full licensing details.
-"""Unit tests for the :mod:`iris.ugrid.metadata` package."""
-
-from __future__ import annotations
